added stripe checkout, deleting laravel cashier

+++ database/migrations/2023_08_25_213126_table_jewelry_products.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('jewelry_products', function (Blueprint $table) {
             $table->id();
             $table->string('nom');
-            $table->double('prix', 10, 2);
+            $table->decimal('prix', 10, 2);
             $table->string('slug')->unique();
 
             $table->string('type');
